+++ tests/testthat.R
@@ -9,4 +9,10 @@
 library(testthat)
 library(nasapower)
 
-test_check("nasapower")
+if (Sys.getenv("NOT_CRAN") == "true") {
+  # like global skip_on_cran
+  # https://github.com/r-lib/testthat/issues/144#issuecomment-396902791
+  # according to https://github.com/hadley/testthat/issues/144
+  Sys.setenv("R_TESTS" = "")
+  test_check("nasapower")
+}