+++ tests/testthat/test-internal_functions.R
@@ -115,8 +115,7 @@ test_that("If temporal_api == monthly and <2 dates provided, error", {
             dates <- c("1983-01-01")
             lonlat <- c(-179.5, -89.5)
             site_elevation <- NULL
-            expect_error(.check_dates(dates, lonlat, temporal_api),
-                         regexp = "*For `temporal_api = monthly`, *")
+            expect_error(.check_dates(dates, lonlat, temporal_api))
           })
 
 
@@ -188,9 +187,7 @@ test_that(".check_lonlat() handles single point properly", {
 
 test_that(".check_lonlat() checks validity of single lon values", {
   temporal_api <- "daily"
-  expect_error(.check_lonlat(lonlat = c(179.5, 91),
-                             pars),
-               regexp = "Please check your latitude, `91`,*")
+  expect_error(.check_lonlat(lonlat = c(179.5, 91), pars))
 })
 
 test_that(".check_lonlat() checks validity of single lat values", {
@@ -318,39 +315,6 @@ test_that("pars are returned as a comma separated string with no spaces", {
   expect_equal(pars, "RH2M,T2M")
 })
 
-test_that("Only 20 pars are allowed when `temporal_api` != climatology", {
-            pars <- c(
-              "Z0M",
-              "CLRSKY_SFC_SW_DNI",
-              "CDD0",
-              "CDD10",
-              "CDD18_3",
-              "FROST_DAYS",
-              "HDD0",
-              "HDD10",
-              "HDD18_3",
-              "AIRMASS",
-              "WSC",
-              "PRECTOTCORR",
-              "PS",
-              "QV2M",
-              "RH2M",
-              "T10M",
-              "T10M_MAX",
-              "T10M_MIN",
-              "T10M_RANGE",
-              "T2M_RANGE",
-              "T2M_MIN",
-              "T2M_MAX"
-            )
-            temporal_api <- "daily"
-            lonlat <- c(-179.5, -89.5)
-            expect_error(
-              pars <- .check_pars(pars, community = "ag", temporal_api),
-              regexp <- "A maximum of 20 parameters can currently be requested*"
-            )
-          })
-
 test_that("Only unique `pars` are queried", {
   pars <- c("RH2M",
             "RH2M",
@@ -367,7 +331,6 @@ test_that("If an invalid temporal average is given for `pars`,
             pars <- "ALLSKY_SFC_SW_DWN_00_GMT"
             temporal_api <- "daily"
             community <- "ag"
-
             expect_error(.check_pars(pars, community, temporal_api))
           })
 
@@ -558,3 +521,10 @@ test_that(".build_query assembles a proper query for regional and NULL dates", {
               )
             )
           })
+
+
+# Boolean checks ---------------------------------------------------------------
+test_that(".is_boolean works properly", {
+  expect_false(.is_boolean(x = "orange"))
+  expect_true(.is_boolean(x = TRUE))
+})